+++ apps/trackers/jira/constants.py
@@ -11,8 +11,10 @@ JIRA_INTERNAL_SECURITY_LEVEL_NAME = get_env(
     "JIRA_INTERNAL_SECURITY_LEVEL_NAME", default="Red Hat Employee"
 )
 
-# Translate additional fields as defined in product definitions to the actual Jira field
+# Translate fields as defined in product definitions to the actual Jira field
 PS_ADDITIONAL_FIELD_TO_JIRA = {
     "fixVersions": "fixVersions",
     "release_blocker": "customfield_12319743",
+    "target_release": "customfield_12311240",
+    "target_version": "customfield_12319940",
 }